Move the Java package of embulk-guess-csv_all_strings to org.embulk.guess.csv_all_strings
