+++ embulk-guess-csv_all_strings/src/test/java/org/embulk/standards/guess/TestCsvAllStringsGuessPlugin.java
@@ -27,8 +27,8 @@ import org.embulk.config.ConfigSource;
 import org.embulk.config.DataSource;
 import org.embulk.formatter.csv.CsvFormatterPlugin;
 import org.embulk.guess.bzip2.Bzip2GuessPlugin;
-import org.embulk.guess.csv.CsvAllStringsGuessPlugin;
 import org.embulk.guess.csv.CsvGuessPlugin;
+import org.embulk.guess.csv_all_strings.CsvAllStringsGuessPlugin;
 import org.embulk.guess.gzip.GzipGuessPlugin;
 import org.embulk.guess.json.JsonGuessPlugin;
 import org.embulk.input.file.LocalFileInputPlugin;
